assignment 13th almost done.

+++ lesson13/Lesson13/Lesson13/Program.cs
@@ -8,9 +8,10 @@ namespace Lesson13
 {
     class Program
     {
-        private static string dir1Path = @"D:\dir1";
-        private static string dir2Path = @"D:\dir2";
-        private static int i = 1;
+        private static string dir1Path = @"D:\dev\Amdaris_lessons\dir1";
+        //@"D:\dir1";
+        private static string dir2Path = @"D:\dev\Amdaris_lessons\dir2";
+            //@"D:\dir2";
         static void Main(string[] args)
         {
             FileSystemWatcher watcher = new FileSystemWatcher();
@@ -20,7 +21,7 @@ namespace Lesson13
 
             try
             {
-                watcher.Path = @"D:\dir1";
+                watcher.Path = dir1Path;
             } 
             catch (Exception e)
             {
@@ -42,21 +43,27 @@ namespace Lesson13
             Console.WriteLine($"Press \"q\" to quit app.");
             while (Console.Read() != 'q') ;
         }
+
         private static void OnCreated(object sender, FileSystemEventArgs e)
         {
-
+            //get name of created file first, not path. Example: text.txt
             var fileName = Path.GetFileName(e.FullPath);
+            //combine created file name with second directory path
+            //i.e. dir2 + fileName. newFilePath contains full path to new file in dir2
+            //which will be created
             var newFilePath = Path.Combine(dir2Path, fileName);
-
-                if (!File.Exists(newFilePath))
-                {
-                    Console.WriteLine($"File with path {newFilePath} doesn't exists in {dir2Path}");
-                    CreateFileInDir2(newFilePath, e.FullPath);
-                }
+            //create file in second directory if it does not exist there.
+            if (!File.Exists(newFilePath))
+            {
+                Console.WriteLine($"File with path {fileName} doesn't exists in {dir2Path}");
+                Console.WriteLine($"Creating file \"{newFilePath}\"");
+                CreateFileInDir2(newFilePath, e.FullPath);
+            }
         }
 
         private static void OnRenamed(object sender, RenamedEventArgs e)
         {
+            //retrieve old and new paths
             var oldPath = Path.GetFileName(e.OldFullPath);
             var newPath = Path.GetFileName(e.FullPath);
             RenameFileInDir2(oldPath, newPath);
@@ -64,47 +71,64 @@ namespace Lesson13
 
         static void OnChanged(object sender, FileSystemEventArgs e)
         {
+            //retrieve name of changed file
             var fileName = Path.GetFileName(e.FullPath);
+            //we wait until changed file is locked to prevent exception
+            while (IsFileLocked(e.FullPath)) ;
+            //now we can go further
+            Console.WriteLine($"File {fileName} is unlocked.");
+            //compute hash for changed file from dir1
             var dir1FileHash = CheckMD5(e.FullPath);
+            //create path for changed file in dir2 and check if such file exists
             var dir2FilePath = Path.Combine(dir2Path, fileName);
             if (File.Exists(dir2FilePath))
             {
+                //now we can compute hash for equivalent file in dir2
                 var dir2FileHash = CheckMD5(dir2FilePath);
+                //and check if file has been actually changed
                 if (!dir1FileHash.Equals(dir2FileHash))
                 {
+                    //delete file with old content and recreate it
                     File.Delete(dir2FilePath);
                     CreateFileInDir2(dir2FilePath, e.FullPath);
                 } 
                 else
                 {
-                    Console.WriteLine("Nothing to change.");
+                    Console.WriteLine($"Nothing to change. File \"{dir2FilePath}\" is up to date.");
                 }
             } 
             else
             {
-                Console.WriteLine($"Could not find file {fileName}");
+                Console.WriteLine($"Could not find file \"{fileName}\"");
             }
         }
 
         private static void OnDeleted(object sender, FileSystemEventArgs e)
         {
+            //get deleted file path
             var fileName = Path.GetFileName(e.FullPath);
-            var fileToDelPath = Path.Combine(dir2Path, fileName);
-            if (File.Exists(fileToDelPath))
+            //retrieve path of equivalent file from dir2 and check if it exists
+            var fileInDir2ToDelPath = Path.Combine(dir2Path, fileName);
+            if (File.Exists(fileInDir2ToDelPath))
             {
-                File.Delete(fileToDelPath);
+                File.Delete(fileInDir2ToDelPath);
+                Console.WriteLine($"File \"{fileName}\" has successfully been deleted in {dir2Path}.");
             } 
             else
             {
-                Console.WriteLine($"Could not find file {fileName}");
+                Console.WriteLine($"Could not find file \"{fileName}\" in {dir2Path}");
             }
         }
 
-        private static async void CreateFileInDir2(string path, string srcPath)
+        private static async void CreateFileInDir2(string destPath, string srcPath)
         {
+            //destPath is the full path of a new file in dir2 which will be created here.
+            //srcPath is the full path of a created file in dir2
+            //wait until file is locked by another process
+            while (IsFileLocked(srcPath)) ;
             using var inStream = new FileStream(srcPath, FileMode.Open);
             using var reader = new StreamReader(inStream);
-            using var outStream = new FileStream(path, FileMode.Create);
+            using var outStream = new FileStream(destPath, FileMode.Create);
             using var writer = new StreamWriter(outStream);
             while (!reader.EndOfStream)
             {
@@ -116,7 +140,13 @@ namespace Lesson13
         {
             var newFileName = Path.GetFileName(newPath);
             var oldFileName = Path.GetFileName(oldPath);
-            File.Move(Path.Combine(dir2Path, oldFileName), Path.Combine(dir2Path, newFileName));
+            //additional checks for existing files in dir2
+            if(File.Exists(Path.Combine(dir2Path, oldFileName)) &&
+                !File.Exists(Path.Combine(dir2Path, newFileName)))
+            {
+                File.Move(Path.Combine(dir2Path, oldFileName), Path.Combine(dir2Path, newFileName));
+                Console.WriteLine($"\"{Path.Combine(dir2Path, oldFileName)}\" renamed to \"{Path.Combine(dir2Path, newFileName)}\"");
+            }
         }
 
         private static string CheckMD5(string filename)
@@ -124,6 +154,20 @@ namespace Lesson13
             using var md5 = MD5.Create();
             using var stream = File.OpenRead(filename);
             return Encoding.Default.GetString(md5.ComputeHash(stream));
+        }
+
+        private static bool IsFileLocked(string path)
+        {
+            try
+            {
+                using FileStream stream = File.Open(path, FileMode.Open, FileAccess.Read, FileShare.None);
+                stream.Close();
+            }
+            catch (Exception)
+            {
+                return true;
+            }
+            return false;
         }
     }
 }
